Move direct ufunc_helper creation functions to their own file.
